requirements.txt
Django==1.9.5
djangorestframework==3.3.3
djangorestframework_jwt==1.8.0
django-cors-headers==1.1.0
PyJWT==1.4.0
cryptography==1.3.1
psycopg2==2.6.1
scrypt==0.7.1
websockets==3.1
requests==2.10.0
